requirements.txt
tensorflow==2.7
protobuf
easyocr
opencv-python==4.5.4.60
gdown
pillow
absl-py
six
tqdm